requirements.txt
azure-ai-formrecognizer==3.3.0
python-dotenv==1.0.0
pyairtable==2.1.0
streamlit==1.27.0
pillow>=7.1.0,<10.0.0
pandas==2.1.0 